Implemented iterator and delete in CAS::Simple, and fixed windows line ending bug
iterator() and delete() were previously {...}, and it turns out that
syntax was new with perl 5.12, so I just implemented them.

There was also a bug reported which appears to be a windows line
ending error, so I added code to account for that.

Heavily updated documentation, too, adding lots of doc-links.

Bumped version number of Simple to 0.0200, as well as whole package.