Option to use full hash as filename in ::Simple
Previously, the filename would be whatever characters didn't get
used for the directory names.  However, sometimes it is nice to
have the full hash in the filename, such as for iterating a single
directory and knowing what files you are working on, or rearranging
the organization of existing files.

This broke the _fanout_regex, but that was a bad API anyway and
I replaced it with ->path_parts_for_hash and ->path_for_hash