Overhaul of add-file behavior and new move feature
The previous design for adding files (or scalars) would all pass
through the stream-adding API, which would write a new file into
the CAS before deciding whether it needed added.  But, it seems
likely that reading a file twice (once to hash, once to copy)
each time a file is new would be better than writing an unneeded
file each time a file exists.  So, I changed the design so that
for scalars and files as input, it hashes first, checks for
existence, and then only writes a file if needed.  In the case of
a real source file and the 'hardlink' feature, the original file
gets linked instead of copied, and if the hash was already known,
it never gets read either, resulting in optimum performance.

I also added a new 'move' feature, that is similar to 'hardlink'
but always removes the source file, even when the file already
existed in the CAS.  It refuses to remove from other CAS instances
though.

I also now remove the write permission bits from the files added
to CAS::Simple.  And, for the 'move' and 'hardlink' features I
ensure that the correct user owns the resulting files, and that
the write permission bits are similarly removed.