Record the whole version hierarchy into Simple's storage dir
The DataStore::CAS::Simple implementation wrote out the version
of itself into a config file in the storage directory, but if it
were subclassed, it would write out only the derived module's
version, and the derived module might not even have one.

Now it warns about missing versions, and also writes a line for
each module in the hierarchy.  (this was already intended, but
I had not finished implementing it before)