Standardize on attribute name 'cas' rather than 'store'
I decided it was a potential for confusion to use the name
'store' since it is both a noun and a verb.  The name of the
main module is 'CAS' so it makes sense to have the attribute
that points to it use the same name.

This also makes the cas attribute of the handle public, where
previously it was private.